backend/requirements.txt
python-dotenv
ipykernel
jupyterlab
jupyterlab-git
openai
openai-agents
pymatgen
numpy<2
git+https://github.com/xraypy/xraylarch.git
ipympl
anywidget
nbformat>=4.2.0
boto3>=1.35.49
pytest>=7.2.1
requests>=2.28.2
fastapi
mp_api